+++ src/decorators/InterfaceType.ts
@@ -1,8 +1,8 @@
 import { getMetadataStorage } from "../metadata/getMetadataStorage";
 import { getNameDecoratorParams } from "../helpers/decorators";
-import { DescriptionOptions, AbstractClassOptions } from "./types";
+import { DescriptionOptions, AbstractClassOptions, ResolveTypeOptions } from "./types";
 
-export type InterfaceOptions = DescriptionOptions & AbstractClassOptions;
+export type InterfaceOptions = DescriptionOptions & AbstractClassOptions & ResolveTypeOptions;
 
 export function InterfaceType(): ClassDecorator;
 export function InterfaceType(options: InterfaceOptions): ClassDecorator;
@@ -16,8 +16,7 @@ export function InterfaceType(
     getMetadataStorage().collectInterfaceMetadata({
       name: name || target.name,
       target,
-      description: options.description,
-      isAbstract: options.isAbstract,
+      ...options,
     });
   };
 }