+++ src/metadata/definitions/union-metadata.ts
@@ -1,9 +1,10 @@
-import { ClassType } from "../../interfaces";
+import { ClassType, TypeResolver } from "../../interfaces";
 
 export interface UnionMetadata {
-  types: ClassType[];
+  classTypes: ClassType[];
   name: string;
   description?: string;
+  resolveType?: TypeResolver<any, any>;
 }
 export interface UnionMetadataWithSymbol extends UnionMetadata {
   symbol: symbol;