+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 <!-- here goes all the unreleased changes descriptions -->
 ### Features
+- add support for defining `resolveType` function for interfaces and unions (#319)
 - add postinstall script for printing info on console about supporting the project
 - add support for setting default nullability for fields and return types (#297)
 - add `skipCheck` option in `buildSchema` to disable checking the correctness of a schema